+++ src/main/java/tech/bonda/PaymentBackEnd/service/TransactionService/TransactionServiceImpl.java
@@ -1,5 +1,6 @@
 package tech.bonda.PaymentBackEnd.service.TransactionService;
 
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import tech.bonda.PaymentBackEnd.entities.transaction.Transaction;
@@ -9,8 +10,12 @@ import java.util.List;
 
 @Service
 public class TransactionServiceImpl implements TransactionService {
+    private final TransactionRepository transactionRepository;
+
     @Autowired
-    private TransactionRepository transactionRepository;
+    public TransactionServiceImpl(TransactionRepository transactionRepository) {
+        this.transactionRepository = transactionRepository;
+    }
 
     @Override
     public Transaction saveTransaction(Transaction transaction) {
@@ -24,23 +29,36 @@ public class TransactionServiceImpl implements TransactionService {
 
     @Override
     public Transaction getTransactionById(long id) {
-        return transactionRepository.findById(id).orElse(null);
+        return transactionRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Transaction not found for ID: " + id));
     }
 
     @Override
     public Transaction updateTransaction(long id, Transaction transaction) {
         Transaction existingTransaction = transactionRepository.findById(id).orElse(null);
-        existingTransaction.setAmount(transaction.getAmount());
-        existingTransaction.setSenderIban(transaction.getSenderIban());
-        existingTransaction.setReceiverIban(transaction.getReceiverIban());
-        existingTransaction.setTimestamp(transaction.getTimestamp());
-        return transactionRepository.save(existingTransaction);
+        if (existingTransaction != null)
+        {
+            existingTransaction.setAmount(transaction.getAmount());
+            existingTransaction.setSenderIban(transaction.getSenderIban());
+            existingTransaction.setReceiverIban(transaction.getReceiverIban());
+            existingTransaction.setTimestamp(transaction.getTimestamp());
+            return transactionRepository.save(existingTransaction);
+        }
+        return null;
     }
 
     @Override
-    public void deleteTransaction(long id) {
-        transactionRepository.deleteById(id);
+    public boolean deleteTransaction(long id) {
+        if (transactionRepository.existsById(id))
+        {
+            transactionRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
-
+    @Override
+    public List<Transaction> getTransactionsByCardId(long cardId) {
+        return transactionRepository.findTransactionsBySenderOrReceiverCardId(cardId);
+    }
 }
+
